+++ packages/native-core/lib/recipe/parameters.js
@@ -81,6 +81,7 @@ exports._setParameters = function _setParameters() {
   };
 
   this._margin = Object.assign({}, this.default.pageMargin);
+  this.contextState = "idle";
 
   // Object.assign(this, options);
 };